refactor: extract shared bootstrapNgModule + compileWith test helpers
