+++ src/compiler/__tests__/class-directives.test.ts
@@ -29,44 +29,10 @@
 
 import { beforeEach, describe, expect, it } from 'vitest';
 
-import { $CompileProvider } from '@compiler/compile-provider';
-import type { Attributes, CompileService, DirectiveFactory, DirectiveFactoryReturn } from '@compiler/directive-types';
+import type { Attributes, DirectiveFactory, DirectiveFactoryReturn } from '@compiler/directive-types';
 import { Scope } from '@core/index';
-import { createInjector } from '@di/injector';
-import { createModule, resetRegistry } from '@di/module';
-import { $FilterProvider } from '@filter/filter-provider';
-import { $InterpolateProvider } from '@interpolate/interpolate-provider';
-import { $SceDelegateProvider } from '@sce/sce-delegate-provider';
-import { $SceProvider } from '@sce/sce-provider';
-import { createTemplateCache } from '@template/template-cache';
-import { createTemplateRequest } from '@template/template-request';
-import type { TemplateCacheService, TemplateRequestFn } from '@template/template-types';
 
-function bootstrapNgModule(): void {
-  resetRegistry();
-  createModule('ng', [])
-    .factory('$exceptionHandler', [() => () => undefined])
-    .provider('$sceDelegate', $SceDelegateProvider)
-    .provider('$sce', $SceProvider)
-    .provider('$interpolate', $InterpolateProvider)
-    .provider('$filter', ['$provide', $FilterProvider])
-    .factory('$templateCache', [() => createTemplateCache()])
-    .factory('$templateRequest', [
-      '$templateCache',
-      (cache: TemplateCacheService): TemplateRequestFn => createTemplateRequest({ cache }),
-    ])
-    .provider('$compile', ['$provide', $CompileProvider]);
-}
-
-function compileWith(register: ($cp: $CompileProvider) => void): CompileService {
-  const appModule = createModule('app', ['ng']).config([
-    '$compileProvider',
-    ($cp: $CompileProvider) => {
-      register($cp);
-    },
-  ]);
-  return createInjector([appModule]).get('$compile');
-}
+import { bootstrapNgModule, compileWith } from './test-helpers';
 
 function ddoFactory(returnValue: DirectiveFactoryReturn): DirectiveFactory {
   return [() => returnValue] as DirectiveFactory;
