+++ src/compiler/__tests__/multiple-directives.test.ts
@@ -32,30 +32,9 @@ import { $CompileProvider } from '@compiler/compile-provider';
 import type { CompileService, Directive, DirectiveFactory, DirectiveFactoryReturn } from '@compiler/directive-types';
 import { Scope } from '@core/index';
 import { createInjector } from '@di/injector';
-import { createModule, resetRegistry } from '@di/module';
-import { $FilterProvider } from '@filter/filter-provider';
-import { $InterpolateProvider } from '@interpolate/interpolate-provider';
-import { $SceDelegateProvider } from '@sce/sce-delegate-provider';
-import { $SceProvider } from '@sce/sce-provider';
-import { createTemplateCache } from '@template/template-cache';
-import { createTemplateRequest } from '@template/template-request';
-import type { TemplateCacheService, TemplateRequestFn } from '@template/template-types';
+import { createModule } from '@di/module';
 
-function bootstrapNgModule(): void {
-  resetRegistry();
-  createModule('ng', [])
-    .factory('$exceptionHandler', [() => () => undefined])
-    .provider('$sceDelegate', $SceDelegateProvider)
-    .provider('$sce', $SceProvider)
-    .provider('$interpolate', $InterpolateProvider)
-    .provider('$filter', ['$provide', $FilterProvider])
-    .factory('$templateCache', [() => createTemplateCache()])
-    .factory('$templateRequest', [
-      '$templateCache',
-      (cache: TemplateCacheService): TemplateRequestFn => createTemplateRequest({ cache }),
-    ])
-    .provider('$compile', ['$provide', $CompileProvider]);
-}
+import { bootstrapNgModule, compileWith } from './test-helpers';
 
 function buildInjector(register: ($cp: $CompileProvider) => void) {
   const appModule = createModule('app', ['ng']).config([
@@ -65,10 +44,6 @@ function buildInjector(register: ($cp: $CompileProvider) => void) {
     },
   ]);
   return createInjector([appModule]);
-}
-
-function compileWith(register: ($cp: $CompileProvider) => void): CompileService {
-  return buildInjector(register).get<CompileService>('$compile');
 }
 
 function ddoFactory(returnValue: DirectiveFactoryReturn): DirectiveFactory {
